Added a column containing the genomic positions to the estimated-regime-probabilities (CSV) file.
Reason: I am unable to reproduce the bug related to the incorrect number of rows in the above-mentioned CSV file on my own machine. Adding a column with the genomic positions should thus at least make it clear to which genomic positions the estimated regime probabilities correspond. At worst, we are thus simply missing the estimated regimes at the very start or very end of the chromosome.